Pull out some functions into utilities
