+++ umbra-js/index.js
@@ -1,7 +1,9 @@
 const KeyPair = require('./classes/KeyPair');
 const RandomNumber = require('./classes/RandomNumber');
+const utils = require('./utils/utils');
 
 module.exports = {
   KeyPair,
   RandomNumber,
+  utils,
 };
